requirements.txt
absl-py==0.14.1
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
cmake==3.26.1
comm==0.1.3
cycler==0.10.0
debugpy==1.6.6
decorator==5.1.0
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.7
flatbuffers==1.12
fqdn==1.5.1
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
fvcore==0.1.5.post20220212
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
huggingface-hub==0.13.3
idna==3.3
importlib-metadata==6.1.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
librosa==0.8.1
lightning-utilities==0.8.0
lit==16.0.0
llvmlite==0.37.0
Markdown==3.3.4
MarkupSafe==2.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==5.2.0
nbclassic==0.5.4
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.54.1
numpy==1.20.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pooch==1.5.2
portalocker==2.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.18.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.14.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.9.0
pytorch-lightning-bolts @ git+https://github.com/PytorchLightning/pytorch-lightning-bolts.git@1bec12200ff60b9b58277b40d9232933e1aa9dbe
pytz==2021.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
SoundFile==0.10.3.post1
soupsieve==2.4
stack-data==0.6.2
sympy==1.11.1
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.17.1
threadpoolctl==3.0.0
timm==0.6.13
tinycss2==1.2.1
torch==1.13.1+cu116
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.62.3
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.0.2
wget==3.2
widgetsnbextension==4.0.7
wrapt==1.12.1
yacs==0.1.8
yarl==1.7.2
zipp==3.15.0
